extract common types to types.ts
